requirements.txt
camelot-py==0.10.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
decorator==5.1.1
Elixir==0.7.1
et-xmlfile==1.1.0
ghostscript==0.7
importlib-metadata==4.11.4
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.21.6
opencv-python==4.5.5.64
openpyxl==3.0.10
pandas==1.3.5
pbr==5.9.0
pdfminer.six==20220524
pdftopng==0.2.3
pycparser==2.21
PyPDF2==2.0.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.4.2
tabulate==0.8.9
Tempita==0.5.2
typing_extensions==4.2.0
xlrd==0.7.1
xlwt==0.7.2
zipp==3.8.0
